:white_check_mark: refactor: Divide Member Fixture (NAILCASE-96)
- 통합 테스트와 단위테스트 분리를 위해 fixture 분리

resoling: [NAILCASE-96](https://nailcase.atlassian.net/browse/NAILCASE-96), [NAILCASE-144](https://nailcase.atlassian.net/browse/NAILCASE-144)
refs: [NAILCASE-94](https://nailcase.atlassian.net/browse/NAILCASE-94),
[NAILCASE-128](https://nailcase.atlassian.net/browse/NAILCASE-128)
